refactor: refactor directory structure, file names and add import aliases (#13)
