+++ packages/core/src/bsky/Preferences.ts
