+++ packages/core/src/tsky/Paginator.ts
