+++ packages/core/src/tsky/index.ts
@@ -1,0 +1,2 @@
+export * from './Paginator';
+export * from './TSky';
